Adds auth package
