+++ ads/ads_auth/CONTRIBUTING.md
@@ -1,0 +1,62 @@
+# Contributing to Oracle Accelerated Data Science (ADS)
+
+We welcome your contributions! There are multiple ways to contribute.
+
+## Opening issues
+
+For bugs or enhancement requests, file a GitHub issue unless it's
+security related. When filing a bug, remember that the better written the bug is,
+the more likely it is to be fixed. If you think you've found a security
+vulnerability, don't raise a GitHub issue and follow the instructions in our
+[security policy](./SECURITY.md).
+
+## Contributing code
+
+We welcome your code contributions. Before submitting code using a pull request,
+you must sign the [Oracle Contributor Agreement](https://oca.opensource.oracle.com) (OCA) and
+your commits need to include the following line using the name and e-mail
+address you used to sign the OCA:
+
+```text
+Signed-off-by: Your Name <you@example.org>
+```
+
+This can be automatically added to pull requests by committing with `--sign-off`
+or `-s`, for example:
+
+```text
+git commit --signoff
+```
+
+Only pull requests from committers that can be verified as having signed the OCA
+are accepted.
+
+## Branching strategy
+
+Use GitHub flow branching strategy. With the GitHub flow, there are only 2 branches: main and feature.
+
+1. Every change that is worked on is branched directly off of main into a feature branch.
+2. Once a feature is ready it is tested on the feature branch and the code is reviewed before being merged to main.
+
+Hotfixes, bugfixes are treated the same as feature branches.
+
+## Pull request process
+
+1. Ensure there is an issue created to track and discuss the fix or enhancement
+   you intend to submit.
+2. Fork this repository.
+3. Create a branch in your fork to implement the changes. We recommend using
+   the issue number as part of your branch name, for example `1234-fixes`.
+4. Ensure that any documentation is updated with the changes that are required
+   by your change.
+5. Ensure that any samples are updated if the base image has been changed.
+6. Submit the pull request. *Don't leave the pull request blank*. Explain exactly
+   what your changes are meant to do and provide simple steps about how to validate
+   your changes. Ensure that you reference the issue you created as well.
+7. We assign the pull request to 2-3 people for review before it is merged.
+
+## Code of conduct
+
+Follow the [Golden Rule](https://en.wikipedia.org/wiki/Golden_Rule). If you'd
+like more specific guidelines, see the
+[Contributor Covenant Code of Conduct](https://www.contributor-covenant.org/version/1/4/code-of-conduct/).
