+++ ads/ads_auth/ads/__init__.py
@@ -1,0 +1,7 @@
+#!/usr/bin/env python
+
+# Copyright (c) 2025 Oracle and/or its affiliates.
+# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/
+from pkgutil import extend_path
+
+__path__ = extend_path(__path__, __name__)
