+++ pyproject.toml
@@ -56,6 +56,7 @@ classifiers = [
 # Copied from install_requires list in setup.py, setup.py got removed in favor of this config file
 dependencies = [
   "PyYAML>=6",                         # pyyaml 5.4 is broken with cython 3
+  "oracle_ads_auth",
   "asteval>=0.9.25",
   "cerberus>=1.3.4",
   "cloudpickle>=1.6.0",
